requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
streamlit
pypdf
python-dotenv
